+++ examples/basic/forms/index.js
@@ -53,7 +53,7 @@ class SimpleUsage extends Component {
   }
 
   headingRef = h1 => this.heading = h1;
-  
+
   render() {
     const { currentModal } = this.state;
 
@@ -79,7 +79,8 @@ class SimpleUsage extends Component {
           contentLabel="modalB"
           isOpen={currentModal == MODAL_B}
           onAfterOpen={this.handleOnAfterOpenModal}
-          onRequestClose={this.toggleModal(MODAL_B)}>
+          onRequestClose={this.toggleModal(MODAL_B)}
+          shouldFocusOnClose={false}>
           <h1 id="heading" ref={headingRef}>This is the modal 2!</h1>
           <div id="fulldescription" tabIndex="0" role="document">
             <p>This is a description of what it does: nothing :)</p>