+++ src/components/ModalPortal.js
@@ -48,7 +48,8 @@ export default class ModalPortal extends Component {
     contentLabel: PropTypes.string,
     aria: PropTypes.object,
     children: PropTypes.node,
-    shouldCloseOnEsc: PropTypes.bool
+    shouldCloseOnEsc: PropTypes.bool,
+    shouldFocusOnClose: PropTypes.bool
   };
 
   constructor(props) {
@@ -135,7 +136,9 @@ export default class ModalPortal extends Component {
   }
 
   afterClose = () => {
-    focusManager.returnFocus();
+    if (this.props.shouldFocusOnClose) {
+      focusManager.returnFocus();
+    }
     focusManager.teardownScopedFocus();
   };
 
@@ -146,7 +149,9 @@ export default class ModalPortal extends Component {
       this.setState({ beforeClose: false });
     } else {
       focusManager.setupScopedFocus(this.node);
-      focusManager.markForFocusLater();
+      if (this.props.shouldFocusOnClose) {
+        focusManager.markForFocusLater();
+      }
       this.setState({ isOpen: true }, () => {
         this.setState({ afterOpen: true });